==> Scripts/4. Selection of the best class of models.Rmd <==
---
title: "Selection of the best class of models"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Loading libraries
```{r}
library('lubridate') # Library for date management
library('ranger') # Library for Random Forest
library('caret') # Library for dummy variables and RMSE
library('xgboost') # Library for Xgboost
library('rpart') # Library for Classfication and Regression Trees
library("glmnet") # Library for LASSO and Ridge regression
```


# Modelisation

## Random Forest :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Model training
rf <- ranger(ecart_sais~., data = visiteurs[, -1])

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(rf, visiteurs_test)$predictions
RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais) # RMSE: 267.1256

# Mean pourcentage error of the prediction
mean(abs(visiteurs_test$predict - visiteurs_test$ecart_sais)*100/(visiteurs_test$ecart_sais + visiteurs_test$saison))

# Plot of the real attendance against predicted for 2017
ggplot(visiteurs_test[year(visiteurs_test$DATE) == 2017,]) + 
    geom_line(aes(x = DATE, y = predict + saison), col = "red") + 
    geom_line(aes(x = DATE, y = ecart_sais + saison)) + theme_bw()

visiteurs_test$predict <- NULL
```

## XGBoost :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Factor variables to dummy variables
DATE <- visiteurs_test$DATE
dmy <- dummyVars( ~. , data = visiteurs[,-1], fullRank = T)
visiteurs <- data.frame(predict(dmy, newdata = visiteurs))
dmy <- dummyVars( ~. , data = visiteurs_test[,-1], fullRank = T)
visiteurs_test <- data.frame(predict(dmy, newdata = visiteurs_test))
visiteurs$descr_generalePas.d.obs <- NULL # This doesn't occur in the test set 

# Data in the right format
dtrain <- xgb.DMatrix(data = as.matrix(visiteurs[,-33]), label=as.matrix(visiteurs$ecart_sais))
dtest <- xgb.DMatrix(data = as.matrix(visiteurs_test[,-33]), label=as.matrix(visiteurs_test$ecart_sais))
watchlist <- list(train=dtrain, test=dtest)

# Paremeters
params <- list(eta = 0.01, gamma = 10, max_depth = 3, subsample = .7, colsample_bytree = .7, min_child_weight = 1)

# Model training 
xgb <- xgb.train(params = params, data = dtrain, watchlist=watchlist, verbose=2, nthread=4, nrounds = 5000, early_stopping_rounds = 150)

# Progression of RMSE with the number of trees 
ggplot(xgb$evaluation_log) + geom_line(aes(x = iter, y= test_rmse)) + geom_line(aes(x = iter, y= train_rmse), color = "red") + theme_bw()

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(xgb, dtest)
RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais) # RMSE: 250.3095
visiteurs_test$predict <- NULL
```

## CART :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Model training 
arbre <- rpart(ecart_sais~., data = visiteurs[,-1], cp = 0.0008, model=TRUE)

# Quickly choose a good complexity parameter and plot the tree
plotcp(arbre)
rpart.plot::rpart.plot(arbre)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(arbre, visiteurs_test)
RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais)
visiteurs_test$predict <- NULL
```

## Simple linear regression :

### Linear regression
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Model training 
lm <- lm(ecart_sais ~ ., data = visiteurs)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lm, visiteurs_test)
caret::RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais) # RMSE: 312.459
```

### LASSO 
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Factor variables to dummy variables
DATE <- visiteurs_test$DATE
dmy <- dummyVars( ~. , data = visiteurs[,-1], fullRank = T)
visiteurs <- data.frame(predict(dmy, newdata = visiteurs))
dmy <- dummyVars( ~. , data = visiteurs_test[,-1], fullRank = T)
visiteurs_test <- data.frame(predict(dmy, newdata = visiteurs_test))
visiteurs$descr_generalePas.d.obs <- NULL

# Model training 
lasso.cv <- cv.glmnet(x = as.matrix(visiteurs[,-33]), y = visiteurs$ecart_sais, family = "gaussian", nfolds = 5, alpha = 1)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lasso.cv, newx = as.matrix(visiteurs_test[,-33]), s = "lambda.1se")[,1]
caret::RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais) # RMSE: 313.1293
visiteurs_test$predict <- NULL
```

## Linear regression with interaction variables :

### Linear regression
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Creating all the interaction variables 
DATE <- visiteurs$DATE
DATE_test <- visiteurs_test$DATE
ecart_sais <- visiteurs$ecart_sais
ecart_sais_test <- visiteurs_test$ecart_sais
visiteurs$DATE <- NULL
visiteurs_test$DATE <- NULL
visiteurs_test$ecart_sais <- NULL
visiteurs$ecart_sais <- NULL

visiteurs <- data.frame(model.matrix(~.^2-1, visiteurs))
visiteurs_test <- data.frame(model.matrix(~.^2-1, visiteurs_test))

variables <- names(visiteurs)

visiteurs_test$ecart_sais <- ecart_sais_test
visiteurs$ecart_sais <- ecart_sais

# Iteratively remove variables based on the significance of the test H0: coefficient = 0
# Threshold for stoping the procedure : p_val = 0.01
lm <- lm(ecart_sais ~ ., data = visiteurs[, c(variables, "ecart_sais")])
seuil_signif <- max(summary(lm)$coefficients[,4])

while(seuil_signif > 0.01){
  
  if(sum(is.na(lm$coefficients)) > 0){
    variables <- variables[variables != sample(names(lm$coefficients[is.na(lm$coefficients)]), 1)]
  }
  else{variables <- variables[variables != names(which.max(summary(lm)$coefficients[,4]))]}
  
  lm <- lm(ecart_sais ~ ., data = visiteurs[, c(variables, "ecart_sais")])
  seuil_signif <- max(summary(lm)$coefficients[,4])
  print(length(variables))
  print(seuil_signif)
}
  
lm <- lm(ecart_sais ~ ., data = visiteurs[, variables])
#save.image("Variables regression linéaire.Rdata") 

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lm, visiteurs_test)
caret::RMSE(pred = visiteurs_test$predict, obs = visiteurs_test$ecart_sais) # RMSE : 269.3086
visiteurs_test$predict <- NULL
```

### LASSO :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Creating all the interaction variables 
DATE <- visiteurs$DATE
DATE_test <- visiteurs_test$DATE
ecart_sais <- visiteurs$ecart_sais
ecart_sais_test <- visiteurs_test$ecart_sais
visiteurs$DATE <- NULL
visiteurs_test$DATE <- NULL
visiteurs_test$ecart_sais <- NULL
visiteurs$ecart_sais <- NULL

visiteurs$descr_generale[visiteurs$descr_generale == "Pas d'obs"] <- "Couvert - Soleil"
visiteurs <- data.frame(model.matrix(~.^2-1, visiteurs))
visiteurs_test <- data.frame(model.matrix(~.^2-1, visiteurs_test))

# Model training 
lasso.cv <- cv.glmnet(x = as.matrix(visiteurs), y = ecart_sais, family = "gaussian", nfolds = 5, alpha = 1)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lasso.cv, newx = as.matrix(visiteurs_test), s = "lambda.1se")[,1]
caret::RMSE(pred = visiteurs_test$predict, obs = ecart_sais_test) # RMSE: 274.9008
visiteurs_test$predict <- NULL
```

### Ridge :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Creating all the interaction variables 
DATE <- visiteurs$DATE
DATE_test <- visiteurs_test$DATE
ecart_sais <- visiteurs$ecart_sais
ecart_sais_test <- visiteurs_test$ecart_sais
visiteurs$DATE <- NULL
visiteurs_test$DATE <- NULL
visiteurs_test$ecart_sais <- NULL
visiteurs$ecart_sais <- NULL

visiteurs$descr_generale[visiteurs$descr_generale == "Pas d'obs"] <- "Couvert - Soleil"
visiteurs <- data.frame(model.matrix(~.^2-1, visiteurs))
visiteurs_test <- data.frame(model.matrix(~.^2-1, visiteurs_test))

# Model training 
lasso.cv <- cv.glmnet(x = as.matrix(visiteurs), y = ecart_sais, family = "gaussian", nfolds = 5, alpha = 0)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lasso.cv, newx = as.matrix(visiteurs_test), s = "lambda.1se")[,1]
caret::RMSE(pred = visiteurs_test$predict, obs = ecart_sais_test) # RMSE: 267.2635
visiteurs_test$predict <- NULL
```

### elastic net :
```{r}
# Loading data
load("Donnees pretes pour modelisation.Rdata")

# Creating all the interaction variables 
DATE <- visiteurs$DATE
DATE_test <- visiteurs_test$DATE
ecart_sais <- visiteurs$ecart_sais
ecart_sais_test <- visiteurs_test$ecart_sais
visiteurs$DATE <- NULL
visiteurs_test$DATE <- NULL
visiteurs_test$ecart_sais <- NULL
visiteurs$ecart_sais <- NULL

visiteurs$descr_generale[visiteurs$descr_generale == "Pas d'obs"] <- "Couvert - Soleil"
visiteurs <- data.frame(model.matrix(~.^2-1, visiteurs))
visiteurs_test <- data.frame(model.matrix(~.^2-1, visiteurs_test))

# Model training 
lasso.cv <- cv.glmnet(x = as.matrix(visiteurs), y = ecart_sais, family = "gaussian", nfolds = 5, alpha = .5)

# Make prediction on test set and compute RMSE
visiteurs_test$predict <- predict(lasso.cv, newx = as.matrix(visiteurs_test), s = "lambda.1se")[,1]
caret::RMSE(pred = visiteurs_test$predict, obs = ecart_sais_test) # RMSE: 272.9989
visiteurs_test$predict <- NULL
```


Final results :

| Model                                        | RMSE          |
| -------------------------------------------- |:-------------:|
| Random Forest                                | 267.1256      |
| **Xgboost**                                  | **250.3095**  |
| CART                                         | 312.4196      |
| Linear regression                            | 312.459       |
| LASSO regression                             | 313.1293      |
| Linear regression with interaction variables | 269.3086      |
| LASSO regression with interaction variables  | 274.9008      |
| Ridge regression with interaction variables  | 267.2635      |
| Elastic net with interaction variables       | 272.9989      |
